redshift-publisher-service/requirements.txt
psycopg2-binary==2.9.9
redshift-connector==2.0.917
